Added storeResults function to save data to a csv
Created 3 constants 2 for the data file and output html paths and 1 for
the headers of the data file.
Added storeResults function which takes the results for a single run of
allTests and appends them to the csv or creates a new csv if it doesn't
already exist.
Added csv and html files to gitignore so the outputs from the script
aren't stored.