Added argParser function to parse arguments for paths
Added argParse function which creates a parser expecting path for the
data csv and output html file.
Added main function which runs a single test and updates the results csv
and graphs.
Changed storeResults, readResults and plotGraph to expect a path parameter
instead of using the path contstants, which have been removed.